requirements.txt
# Enterprise RAG Assistant - Dependencies
# Python 3.10+

# Core
python-dotenv>=1.0.0
pyyaml>=6.0

# Document loading
pypdf>=3.0.0
unstructured[pdf]>=0.10.0

# Embeddings & vector store
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4
numpy>=1.24.0

# LangChain
langchain>=0.1.0
langchain-community>=0.0.10
langchain-huggingface>=0.0.1
langchain-text-splitters>=0.0.1

# LLM
transformers>=4.35.0
accelerate>=0.24.0
torch>=2.0.0

# API & UI
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
streamlit>=1.28.0
httpx>=0.25.0

# Dev / tests
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Encoding detection (used by LangChain TextLoader autodetect)
chardet>=5.2.0
